Refactor components into their own folders
This will open space for me to store unit tests
for each component into their folder, as well
as simplify working on the components.

Also set up a helpers.js file to common helpers
to be stored in.
