+++ src/components/CategoryFilter/CategoryFilter.js
@@ -1,0 +1,36 @@
+import React from 'react';
+import CATEGORIES from '../../Categories';
+
+function CategoryFilter({ setCurrentCategory }) {
+  return (
+    <aside>
+      <ul>
+        <li className="category">
+          <button
+            className="btn btn-all-categories"
+            onClick={() => {
+              setCurrentCategory("all");
+            }}
+          >
+            All
+          </button>
+        </li>
+        {CATEGORIES.map((category) => (
+          <li key={category.name} className="category">
+            <button
+              style={{ backgroundColor: category.color }}
+              className="btn btn-categories"
+              onClick={() => {
+                setCurrentCategory(category.name);
+              }}
+            >
+              {category.name}
+            </button>
+          </li>
+        ))}
+      </ul>
+    </aside>
+  );
+}
+
+export default CategoryFilter;
