+++ src/components/Header/Header.js
@@ -1,0 +1,23 @@
+import React from 'react';
+
+function Header({ showForm, setShowForm }) {
+  const appTitle = "Today I Learned";
+
+  return (
+    <header className="header">
+      <div className="logo">
+        <img src="logo.png" alt="The logo for the TiL App" />
+        <h1>{appTitle}</h1>
+      </div>{" "}
+      {/* // Change the value of the state variable */}
+      <button
+        className="btn btn-large btn-share"
+        onClick={() => setShowForm((c) => !c)}
+      >
+        {showForm ? "Close" : "Share a Fact"}
+      </button>
+    </header>
+  );
+}
+
+export default Header;
