+++ src/components/Loader/Loader.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Loader() {
+    return <p className="message">Loading...</p>;
+}
+
+export default Loader;
