+++ src/utils/helpers.js
@@ -1,0 +1,11 @@
+export function isValidHttpUrl(string) {
+    let url;
+  
+    try {
+      url = new URL(string);
+    } catch (_) {
+      return false;
+    }
+  
+    return url.protocol === "http:" || url.protocol === "https:";
+}
